feat:(sh) Change apollo config `eureka.instance.ip-address`
It is recommenbed that if you are useing default built-in registration center,add `-Deureka.instance.ip-address=127.0.0.1` parameter. Because the current IP is used by default,it is currently the internal IP of the container(172.17.0.1/24), which is okay it can not be provided to outside calls. Useing scenarios such as (javasdk/openapi), it is vert useful to verify these functions. Because the `RemoteConfigRepository` in `apollo-client` uses the registration address of the service to send the request. Then get the relavant configuration infomation.

The log is as follows:

Success:
Long polling from http://127.0.0.1:8080/notifications/v2?cluster=default&appId=SampleApp&ip=192.168.12.113&。。。

Fail:
Long polling from http://172.17.0.3:8080/notifications/v2?cluster=default&appId=SampleApp&ip=192.168.12.113&。。。


# Note
You need to rebuild the `apollo-quick-start` container.

`docker build -t nobodyiam/apollo-quick-start . -f ./Dockerfile`